requirements.txt
Django
django-filter
django-extensions
djangorestframework
requests